requirements.txt
python-dotenv==1.0.1
google-generativeai==0.8.2
openmeteo_requests
requests>=2.31.0
matplotlib
numpy
pandas
streamlit
aiohttp>=3.10,<3.11
-e ./sems_portal_api_master